requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
psycopg2-binary
pydantic
pydantic-settings
email-validator
python-jose
passlib
python-multipart
python-dotenv
httpx
jinja2
bcrypt
langchain
openai
beautifulsoup4
robotparser
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
